requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pause==0.3
pyaes==1.6.1
pyasn1==0.6.0
python-dotenv==1.0.1
python-environ==0.4.54
rsa==4.9
Telethon==1.34.0
Werkzeug==3.0.2
